Added bunch of Map stuff to common

+++ KvC-Common/src/main/java/nl/groep4/kvc/common/map/Coordinate.java
@@ -6,10 +6,10 @@ public final class Coordinate implements Serializable, Cloneable {
 
     private static final long serialVersionUID = 12345678987654321L;
 
-    private short x;
-    private short y;
+    private double x;
+    private double y;
 
-    public Coordinate(short x, short y) {
+    public Coordinate(double x, double y) {
 	this.x = x;
 	this.y = y;
     }
@@ -18,27 +18,27 @@ public final class Coordinate implements Serializable, Cloneable {
 	return String.format("%d;%d", x, y);
     }
 
-    public boolean isSameLocation(short x, short y) {
+    public boolean isSameLocation(double x, double y) {
 	return this.x == x && this.y == y;
     }
 
-    public Coordinate add(short x, short y) {
-	return new Coordinate((short) (this.x + x), (short) (this.y + y));
+    public Coordinate add(double x, double y) {
+	return new Coordinate(this.x + x, this.y + y);
     }
 
-    public Coordinate multiply(short x, short y) {
-	return new Coordinate((short) (this.x * x), (short) (this.y * y));
+    public Coordinate multiply(double x, double y) {
+	return new Coordinate(this.x * x, this.y * y);
     }
 
-    public Coordinate subtract(short x, short y) {
-	return new Coordinate((short) (this.x / x), (short) (this.y / y));
+    public Coordinate subtract(double x, double y) {
+	return new Coordinate(this.x / x, this.y / y);
     }
 
-    public short getX() {
+    public double getX() {
 	return this.x;
     }
 
-    public short getY() {
+    public double getY() {
 	return this.y;
     }
